Migrate some language tests to integration tests
