--- subprojects/language-cpp/src/test/groovy/dev/nokee/language/cpp/CppLanguagePluginTest.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2020-2021 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package dev.nokee.language.cpp;
-
-import dev.nokee.language.cpp.internal.plugins.CppLanguageBasePlugin;
-import dev.nokee.language.cpp.internal.plugins.CppLanguagePlugin;
-import lombok.val;
-import org.junit.jupiter.api.Test;
-
-import static dev.nokee.internal.testing.util.ProjectTestUtils.rootProject;
-import static java.util.Collections.singletonMap;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-class CppLanguagePluginTest {
-	@Test
-	void appliesCppLanguageBasePlugin() {
-		val project = rootProject();
-		project.apply(singletonMap("plugin", CppLanguagePlugin.class));
-		assertTrue(project.getPlugins().hasPlugin(CppLanguageBasePlugin.class), "should apply C++ language base plugin");
-	}
-
-//	@Test
-//	void createsCppSourceSetUponDiscoveringSourceSetExtensibleType() {
-//		val project = rootProject();
-//		project.apply(singletonMap("plugin", CppLanguagePlugin.class));
-//		val modelRegistry = project.getExtensions().getByType(ModelRegistry.class);
-//		modelRegistry.register(NodeRegistration.of("test", of(CppSourceSetExtensible.class)));
-//		assertDoesNotThrow(() -> modelRegistry.get("test.cpp", CppSourceSet.class));
-//	}
-}
