+++ subprojects/language-objective-c/language-objective-c.gradle
@@ -1,6 +1,7 @@
 plugins {
 	id 'nokeebuild.java-gradle-plugin'
 	id 'nokeebuild.gradle-plugin-unit-test'
+	id 'nokeebuild.gradle-plugin-integration-test'
 	id 'nokeebuild.gradle-plugin-functional-test'
 	id 'java-test-fixtures'
 	id 'maven-publish'
@@ -10,6 +11,7 @@ dependencies {
 	api project(':languageC')
 	implementation project(':coreScript')
 	implementation project(':coreUtils')
+	implementation project(':runtimeNative')
 	implementation "com.google.guava:guava:${guavaVersion}"
 
 	testImplementation project(':languageC')
@@ -32,6 +34,14 @@ gradlePlugin {
 			id = 'dev.nokee.objective-c-language'
 			implementationClass = 'dev.nokee.language.objectivec.internal.plugins.ObjectiveCLanguagePlugin'
 		}
+	}
+}
+
+integrationTest {
+	dependencies {
+		implementation project(':platformBase')
+		implementation testFixtures(project)
+		implementation testFixtures(project(':coreScript'))
 	}
 }
