+++ subprojects/language-objective-c/src/integrationTest/java/dev/nokee/language/objectivec/ObjectiveCLanguagePluginIntegrationTest.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package dev.nokee.language.objectivec;
+
+import dev.nokee.internal.testing.AbstractPluginTest;
+import dev.nokee.internal.testing.PluginRequirement;
+import org.junit.jupiter.api.Test;
+
+import static dev.nokee.internal.testing.ProjectMatchers.hasPlugin;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+@PluginRequirement.Require(id = "dev.nokee.objective-c-language")
+class ObjectiveCLanguagePluginIntegrationTest extends AbstractPluginTest {
+	@Test
+	void appliesObjectiveCLanguageBasePlugin() {
+		assertThat(project, hasPlugin("dev.nokee.objective-c-language-base"));
+	}
+}
